# docs/swagger.yaml
openapi: 3.0.3
...
info:
  title: Turnito API
  version: 1.0.0
  description: |
    **Turnito** es una plataforma SaaS de gestión de citas multi-tenant.

    Esta API REST permite a negocios gestionar empleados, servicios, horarios
    laborales y citas de forma segura y escalable.

    ## Autenticación
    La API utiliza **JWT Bearer Tokens** emitidos por Supabase Auth.
    Incluye el header `Authorization: Bearer {token}` en cada petición protegida.

    ## Formato de Respuesta Estándar
    ```json
    {
      "success": true,
      "message": "Descripción de la operación",
      "data": { }
    }
    ```

    ## Formato de Error Estándar
    ```json
    {
      "success": false,
      "message": "Descripción del error",
      "errors": [ ]
    }
    ```

    ## Multi-Tenancy
    Turnito soporta múltiples negocios (tenants). Las operaciones de escritura
    sobre recursos de un negocio requieren que el usuario sea propietario o
    empleado activo del negocio correspondiente.
servers:
  - url: http://localhost:3000
    description: Servidor de Desarrollo Local
  - url: https://turnitoapi.onrender.com
    description: Servidor de Producción
...
security:
  - BearerAuth: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT emitido por Supabase Auth.
        Incluir en el header: `Authorization: Bearer {access_token}`

  # ─────────────────────────────────────────────────────────────────────────────
  # Schemas (Modelos de Datos)
  # ─────────────────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operación realizada correctamente.
        data:
          type: object
          nullable: true
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Descripción del error
        errors:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ValidationError'
        stack:
          type: string
          description: Stack trace (solo en desarrollo)
          nullable: true
      required:
        - success
        - message
    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: El formato del correo electrónico es inválido

    # ── Profile ──────────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Error de validación o datos incorrectos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Error de validación en el cuerpo de la solicitud"
            errors:
              - field: email
                message: "El formato del correo electrónico es inválido"
    Unauthorized:
      description: Token no proporcionado, inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Token de autenticación no proporcionado o formato incorrecto"
    Forbidden:
      description: El usuario no tiene permisos para esta acción
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Acceso denegado: solo los administradores pueden realizar esta acción."
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Recurso no encontrado"
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Algo salió mal en el servidor. Por favor, inténtelo de nuevo más tarde."

# ══════════════════════════════════════════════════════════════════════════════
# ENDPOINTS (PATHS)
# ══════════════════════════════════════════════════════════════════════════════
paths:
...
  /upload/avatar:
    post:
      tags: [Uploads]
      summary: Subir imagen de perfil del usuario autenticado
      description: |
        Sube un archivo de imagen para actualizar el avatar del usuario autenticado.
        El archivo se almacena en Supabase Storage dentro del bucket `avatars` bajo la ruta `avatars/{userId}/archivo`.
        
        **Acceso:** Privado – Requiere Bearer Token
      operationId: uploadAvatar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo de imagen a subir (JPEG, PNG o WebP, máx. 5MB)
              required:
                - file
      responses:
        '201':
          description: Imagen subida correctamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          url:
                            type: string
                            format: uri
                            example: "https://bczslyugprpyiohhthkq.supabase.co/storage/v1/object/public/avatars/550e8400-e29b-41d4-a716-446655440000/1687498200000-abcd.jpg"
              example:
                success: true
                message: "Avatar de usuario subido correctamente."
                data:
                  url: "https://bczslyugprpyiohhthkq.supabase.co/storage/v1/object/public/avatars/550e8400-e29b-41d4-a716-446655440000/1687498200000-abcd.jpg"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /upload/logo:
    post:
      tags: [Uploads]
      summary: Subir logo de un negocio
      description: |
        Sube un archivo de imagen para el logo de un negocio específico.
        El usuario autenticado debe tener permisos en el negocio (propietario o rol administrativo).
        El archivo se almacena en Supabase Storage dentro del bucket `business-logos` bajo la ruta `business-logos/{businessId}/archivo`.
        
        **Acceso:** Privado – Requiere Bearer Token + Permisos en el negocio
      operationId: uploadLogo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo de imagen a subir (JPEG, PNG o WebP, máx. 5MB)
                businessId:
                  type: string
                  format: uuid
                  description: UUID del negocio al que pertenece el logo
              required:
                - file
                - businessId
      responses:
        '201':
          description: Logo subido correctamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          url:
                            type: string
                            format: uri
                            example: "https://bczslyugprpyiohhthkq.supabase.co/storage/v1/object/public/business-logos/a1b2c3d4-e5f6-7890-abcd-ef1234567890/1687498200000-xyz.png"
              example:
                success: true
                message: "Logo del negocio subido correctamente."
                data:
                  url: "https://bczslyugprpyiohhthkq.supabase.co/storage/v1/object/public/business-logos/a1b2c3d4-e5f6-7890-abcd-ef1234567890/1687498200000-xyz.png"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'